+++ app/profile/components/index.ts
@@ -1,0 +1,7 @@
+// app/profile/components/index.ts
+
+export { ProfileCard } from "./profile-card";
+export { PrayerJourneyCard } from "./prayer-journey-card";
+export { TheologicalPreferencesCard } from "./theological-preferences-card";
+export { FamilySpaceCard } from "./family-space-card";
+export { PreviousQuestionsCard } from "./previous-questions-card";
